requirements.txt
fastapi
uvicorn[standard]
pydantic
huggingface_hub
cerebras_cloud_sdk
python-dotenv
google-genai
